pyproject.toml
[tool.poetry]
name = "airline-data-engineering-kafka-docker"
version = "0.1.0"
description = ""
authors = ["Bansari Vadgama <bansari.vadgama@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
apache-airflow = "^2.10.2"
cassandra-driver = "^3.29.2"
confluent-kafka = "^2.5.3"
kafka-python = "^2.0.2"
mysql-connector-python = "^9.0.0"
pandas = "^2.1.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
